+++ src/main/java/com/example/capstoneproject/Dto/responses/TransactionResponse.java
@@ -10,9 +10,9 @@ import javax.validation.constraints.NotNull;
 @Getter @Setter
 public class TransactionResponse {
     @NotNull
-    private Long expenditure;
+    private Double expenditure;
 
-    private Long conversionAmount;
+    private Double conversionAmount;
 
     @NotNull
     private Integer userId;
